backend/requirements_full.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
flask==3.0.0
flask-cors==4.0.0
pydantic==2.5.0
kerykeion==4.4.2
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
redis==5.0.8
openai==1.50.2
python-dotenv==1.0.1
sentry-sdk==2.16.0
pytest==8.3.3
httpx==0.27.2
